# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Api Service
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /user_history:
    x-swagger-router-controller: userHistory.controller
    get:
      description: 'Returns history of user messages founded by requested parameters'
      operationId: getUserHistory
      parameters:
        - name: key
          in: query
          description: id or login of user
          required: true
          type: string
        - name: more_than
          in: query
          description: id or date of message
          required: false
          type: string
        - name: less_than
          in: query
          description: id or date of message
          required: false
          type: string

      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/BaseResponse"
...
  /actual_channel_users:
    x-swagger-router-controller: actualChannelUsers.controller
    get:
      description: Return actual users of requested channel
      operationId: getActualChannelUsers
      parameters:
        - name: key
          in: query
          description: id or link of channel
          required: true
          type: string
        - name: date
          in: query
          description: date of channel users state
          required: false
          type: string

      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/BaseResponse"
  # /new_account:
  #   x-swagger-router-controller: account.controller
  #   get:
  #     description: Adding row in account table with certain parameters
  #     operationId: addAccount
  #     parameters:
  #       - name: name
  #         in: query
  #         description: account name
  #         required: true
  #         type: string
  #       - name: login
  #         in: query
  #         description: account login
  #         required: true
  #         type: string
  #       - name: phone
  #         in: query
  #         description: account name
  #         required: true
  #         type: string
  #     responses:
  #       "200":
  #         description: Success
  #         schema:
  #           $ref: "#/definitions/BaseResponse"
...
definitions:
  BaseResponse:
    required:
      - current_timestamp
      - error
      - data
    properties:
      current_timestamp:
        type: string
      error:
        type: 
          - object
          - 'null'
      data:
        type: 
          - object
          - 'null'